+++ src/models/bgp/linkstate.rs
@@ -185,7 +185,7 @@ impl Tlv {
     }
 
     pub fn length(&self) -> u16 {
-        self.value.len() as u16
+        self.value.len().min(u16::MAX as usize) as u16
     }
 }