+++ src/parser/mrt/messages/table_dump_v2/rib_afi_entries.rs
@@ -197,7 +197,7 @@ impl RibEntry {
             }
         }
         let attr_bytes = self.attributes.encode(AsnLength::Bits32);
-        bytes.put_u16(attr_bytes.len() as u16);
+        bytes.put_u16(attr_bytes.len().min(u16::MAX as usize) as u16);
         bytes.extend(attr_bytes);
         bytes.freeze()
     }